% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.outForest}
\alias{is.outForest}
\title{Type Check}
\usage{
is.outForest(x)
}
\arguments{
\item{x}{Any object.}
}
\value{
A logical vector of length one.
}
\description{
Checks if an object inherits class 'outForest'.
}
\examples{
a <- outForest(iris)
is.outForest(a)
is.outForest("a")
}
