% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.outForest}
\alias{print.outForest}
\title{Prints outForest}
\usage{
\method{print}{outForest}(x, ...)
}
\arguments{
\item{x}{A on object of class \code{outForest}.}

\item{...}{Further arguments passed from other methods.}
}
\value{
Invisibly, the input is returned.
}
\description{
Print method for an object of class \code{outForest}.
}
\examples{
x <- outForest(iris)
x
}
