% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.outForest}
\alias{summary.outForest}
\title{Summarizes outForest}
\usage{
\method{summary}{outForest}(object, ...)
}
\arguments{
\item{object}{A on object of class \code{outForest}.}

\item{...}{Further arguments passed from other methods.}
}
\value{
A list of summary statistics.
}
\description{
Summary method for an object of class \code{outForest}. Besides the number of outliers per variables, it shows the worst outliers.
}
\examples{
out <- outForest(iris, seed = 34, verbose = 0)
summary(out)
}
