% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{Data}
\alias{Data}
\alias{Data.default}
\alias{Data.outForest}
\title{Data of outForest}
\usage{
Data(object, ...)

\method{Data}{default}(object, ...)

\method{Data}{outForest}(object, ...)
}
\arguments{
\item{object}{An object of class \code{outForest}.}

\item{...}{Arguments passed from or to other methods.}
}
\value{
A \code{data.frame}.
}
\description{
Extracts data with replaced outliers from object of class \code{outForest}.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Default method not implemented yet.

\item \code{outForest}: Extract data from outForest object.
}}

\examples{
x <- outForest(iris)
head(Data(x))
}
